+++ pages/api/folderPageApi.ts
@@ -4,8 +4,8 @@ import { mapFolderFromLink } from "@/src/util/mapFolderFromLink";
 export const getFolderData = () => instance.get("folders");
 
 export const getLinks = async (folderId: number = 0) => {
-  const folderQuery = folderId === 0 ? "" : `?folderId=${folderId}`;
-  const response = await instance.get(`links${folderQuery}`);
+  const folderQuery = folderId === 0 ? "" : `folders/${folderId}/`;
+  const response = await instance.get(`${folderQuery}links`);
   const data = mapFolderFromLink(response.data?.data.folder);
 
   return data;
