--- pages/api/signinPage.ts
@@ -1,13 +1,0 @@
-import instance from "@/src/util/instance";
-
-export interface FormValue {
-  email: string;
-  password: string;
-  passwordConfirm?: string;
-}
-
-export const postUserInfo = (signinData?: FormValue) =>
-  instance.post("sign-in", signinData);
-
-export const postAccountCheck = (data: FormValue) =>
-  instance.post("sign-up", data);
