+++ src/util/instance.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosInstance } from "axios";
 
 const instance: AxiosInstance = axios.create({
-  baseURL: "https://bootcamp-api.codeit.kr/api/",
+  baseURL: "https://bootcamp-api.codeit.kr/api/linkbrary/v1/",
 });
 
 const getToken = () => {
